+++ samples/react-native/src/App.tsx
@@ -84,7 +84,6 @@ Sentry.init({
       Sentry.reactNativeTracingIntegration({
         // The time to wait in ms until the transaction will be finished, For testing, default is 1000 ms
         idleTimeoutMs: 5_000,
-        traceFetch: false, // Creates duplicate span for axios requests
       }),
       Sentry.httpClientIntegration({
         // These options are effective only in JS.
@@ -100,24 +99,25 @@ Sentry.init({
         maskAllImages: true,
         maskAllVectors: true,
         maskAllText: true,
-        enableExperimentalViewRenderer: true,
+        enableViewRendererV2: true,
         enableFastViewRendering: true,
       }),
       Sentry.appStartIntegration({
         standalone: false,
       }),
       Sentry.reactNativeErrorHandlersIntegration({
-        patchGlobalPromise: Platform.OS === 'ios' && isTurboModuleEnabled()
-          // The global patch doesn't work on iOS with the New Architecture in this Sample app
-          // In
-          ? false
-          : true,
+        patchGlobalPromise:
+          Platform.OS === 'ios' && isTurboModuleEnabled()
+            ? // The global patch doesn't work on iOS with the New Architecture in this Sample app
+              // In
+              false
+            : true,
       }),
       Sentry.feedbackIntegration({
         imagePicker: ImagePicker,
         enableScreenshot: true,
         enableTakeScreenshot: true,
-        styles:{
+        styles: {
           submitButton: {
             backgroundColor: '#6a1b9a',
             paddingVertical: 15,